inventory_shelve/requirement.txt
# Web app
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
WTForms==3.1.2
requests==2.31.0
python-dotenv==1.0.0

# Reconciliation
gspread==5.12.0
oauth2client==4.1.3
